+++ CHANGES.md
@@ -10,3 +10,8 @@ pagurus_game_std
 ----------------
 
 - [FIX] Fix zero division bug during alpha blending where both src and dst alphas are zero
+
+pagurus_android_system
+----------------------
+
+- [CHANGE] Use R8G8B8 as the pixel format instead of R5G6B5